Account time splits FE
